Fix 64-bit build warnings
This fixes two recently introduced 64-bit build warnings.

+++ foo_ui_columns/playlist_tabs.cpp
@@ -271,7 +271,7 @@ LRESULT WINAPI PlaylistTabs::hook(HWND wnd, UINT msg, WPARAM wp, LPARAM lp)
         case WM_CREATE: {
             const auto child_window = reinterpret_cast<HWND>(lp);
             std::array<wchar_t, 128> class_name{};
-            GetClassName(child_window, class_name.data(), class_name.size());
+            GetClassName(child_window, class_name.data(), gsl::narrow<int>(class_name.size()));
 
             if (!wcsncmp(UPDOWN_CLASSW, class_name.data(), class_name.size())) {
                 m_up_down_control_wnd = child_window;
